requirements.txt
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
blinker==1.8.2
certifi==2025.8.3
click==8.1.7
colorama==0.4.6
distro==1.9.0
fastapi==0.116.1
Flask==3.0.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.10.0
MarkupSafe==2.1.5
numpy==2.1.2
openai==1.106.1
pandas==2.2.3
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
starlette==0.47.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.2
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.0.3
